WIP: Vulkan on Linux
Apply after https://chromium-review.googlesource.com/412267.

BUG=angleproject:1668

+++ src/libANGLE/renderer/vulkan/BufferVk.cpp
@@ -55,7 +55,7 @@ gl::Error BufferVk::setData(ContextImpl *context,
 
     // The requirements size is not always equal to the specified API size.
     ASSERT(memoryRequirements.size >= size);
-    mRequiredSize = memoryRequirements.size;
+    mRequiredSize = static_cast<size_t>(memoryRequirements.size);
 
     VkPhysicalDeviceMemoryProperties memoryProperties;
     vkGetPhysicalDeviceMemoryProperties(contextVk->getRenderer()->getPhysicalDevice(),
